Check if gene info exist

+++ download_data.py
@@ -49,32 +49,35 @@ def download_genie(genie, credentials):
 
 
 def download_gene_info(genie):
-    df_genes = pd.read_csv(genie["data_path"] + "data_mutations_extended_13.3-consortium.txt",
-                           sep="\t")
-    df_genes = df_genes["Hugo_Symbol"].drop_duplicates().reset_index(drop=True)
-    df_genes = pd.DataFrame(df_genes)
-    # df_genes["gene_info"] = df_genes["Hugo_Symbol"].apply(lambda x: find_gene_symbol(x))
-    genes = list(df_genes["Hugo_Symbol"].values)
-    total_genes = len(genes)
-    info_genes = []
-    n_gene = 1
-    for gene in genes:
-        log.info(f'Downloading gene {gene}: {n_gene}/{total_genes}')
-        data_gene = find_gene_symbol(gene)
-        info_genes.append(data_gene)
-        time.sleep(0.01)
-        n_gene+=1
-    df_genes["gene_info"] = info_genes
-    df_genes["gene_info"] = df_genes["gene_info"].fillna("NA")
-    df_genes["title"] = df_genes["gene_info"].apply(
-        lambda x: x["title"] if 'title' in x else None)
-    df_genes["status"] = df_genes["gene_info"].apply(
-        lambda x: x["status"] if 'status' in x else None)
-    df_genes["description"] = df_genes["gene_info"].apply(
-        lambda x: x["description"] if 'description' in x else None)
-    cols_to_save = ["Hugo_Symbol", "title", "status", "description"]
-    df_genes[cols_to_save].to_csv(genie["processed_data"] + genie["processed_files"]["gene_info"]["file_name"],
-                                  sep=";", index=False)
+    if os.path.isfile(genie["processed_data"] + genie["processed_files"]["gene_info"]["file_name"]):
+        log.info('Gene info already download, skipping step')
+    else:
+        df_genes = pd.read_csv(genie["data_path"] + "data_mutations_extended_13.3-consortium.txt",
+                               sep="\t")
+        df_genes = df_genes["Hugo_Symbol"].drop_duplicates().reset_index(drop=True)
+        df_genes = pd.DataFrame(df_genes)
+        # df_genes["gene_info"] = df_genes["Hugo_Symbol"].apply(lambda x: find_gene_symbol(x))
+        genes = list(df_genes["Hugo_Symbol"].values)
+        total_genes = len(genes)
+        info_genes = []
+        n_gene = 1
+        for gene in genes:
+            log.info(f'Downloading gene {gene}: {n_gene}/{total_genes}')
+            data_gene = find_gene_symbol(gene)
+            info_genes.append(data_gene)
+            time.sleep(0.01)
+            n_gene+=1
+        df_genes["gene_info"] = info_genes
+        df_genes["gene_info"] = df_genes["gene_info"].fillna("NA")
+        df_genes["title"] = df_genes["gene_info"].apply(
+            lambda x: x["title"] if 'title' in x else None)
+        df_genes["status"] = df_genes["gene_info"].apply(
+            lambda x: x["status"] if 'status' in x else None)
+        df_genes["description"] = df_genes["gene_info"].apply(
+            lambda x: x["description"] if 'description' in x else None)
+        cols_to_save = ["Hugo_Symbol", "title", "status", "description"]
+        df_genes[cols_to_save].to_csv(genie["processed_data"] + genie["processed_files"]["gene_info"]["file_name"],
+                                      sep=";", index=False)
 
 
 def find_gene_symbol(gene):
